refactor: 도메인 모델을 사다리와 참여자로 패키지 분리
