+++ src/main/java/model/ladder/Ladder.java
@@ -1,4 +1,4 @@
-package model;
+package model.ladder;
 
 import java.util.ArrayList;
 import java.util.Collections;
