+++ src/main/java/model/ladder/Width.java
@@ -1,4 +1,4 @@
-package model;
+package model.ladder;
 
 import java.util.Objects;
