+++ src/main/java/model/players/Player.java
@@ -1,4 +1,4 @@
-package model;
+package model.players;
 
 import exception.Message;
