+++ src/test/java/model/HeightTest.java
@@ -1,5 +1,6 @@
 package model;
 
+import model.ladder.Height;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
