+++ src/test/java/model/PlayerTest.java
@@ -1,5 +1,6 @@
 package model;
 
+import model.players.Player;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
