+++ src/test/java/model/PlayersTest.java
@@ -1,6 +1,7 @@
 package model;
 
 import java.util.List;
+import model.players.Players;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
